requirements.txt
astroid==1.6.3
confusable-homoglyphs==3.1.1
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==10.0.1
django-registration==2.4.1
django-tinymce==2.7.0
gunicorn==19.8.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.1.0
psycopg2==2.7.4
pylint==1.8.4
python-decouple==3.1
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
wrapt==1.10.11
